Refactor `TranslatorSv2`
