+++ roles/translator/src/lib/proxy_config.rs
@@ -15,6 +15,42 @@ pub struct ProxyConfig {
     pub upstream_difficulty_config: UpstreamDifficultyConfig,
 }
 
+pub struct TranslatorProxyUpstream {
+    address: String,
+    port: u16,
+    authority_pubkey: Secp256k1PublicKey,
+    difficulty_config: UpstreamDifficultyConfig,
+}
+
+pub struct TranslatorProxyDownstream {
+    address: String,
+    port: u16,
+    difficulty_config: DownstreamDifficultyConfig,
+}
+
+impl ProxyConfig {
+    pub fn new(
+        upstream: TranslatorProxyUpstream,
+        downstream: TranslatorProxyDownstream,
+        max_supported_version: u16,
+        min_supported_version: u16,
+        min_extranonce2_size: u16,
+    ) -> Self {
+        Self {
+            upstream_address: upstream.address,
+            upstream_port: upstream.port,
+            upstream_authority_pubkey: upstream.authority_pubkey,
+            downstream_address: downstream.address,
+            downstream_port: downstream.port,
+            max_supported_version,
+            min_supported_version,
+            min_extranonce2_size,
+            downstream_difficulty_config: downstream.difficulty_config,
+            upstream_difficulty_config: upstream.difficulty_config,
+        }
+    }
+}
+
 #[derive(Debug, Deserialize, Clone)]
 pub struct DownstreamDifficultyConfig {
     pub min_individual_miner_hashrate: f32,
@@ -25,6 +61,21 @@ pub struct DownstreamDifficultyConfig {
     pub timestamp_of_last_update: u64,
 }
 
+impl DownstreamDifficultyConfig {
+    pub fn new(
+        min_individual_miner_hashrate: f32,
+        shares_per_minute: f32,
+        submits_since_last_update: u32,
+        timestamp_of_last_update: u64,
+    ) -> Self {
+        Self {
+            min_individual_miner_hashrate,
+            shares_per_minute,
+            submits_since_last_update,
+            timestamp_of_last_update,
+        }
+    }
+}
 impl PartialEq for DownstreamDifficultyConfig {
     fn eq(&self, other: &Self) -> bool {
         other.min_individual_miner_hashrate.round() as u32
@@ -40,4 +91,20 @@ pub struct UpstreamDifficultyConfig {
     pub timestamp_of_last_update: u64,
     #[serde(default = "bool::default")]
     pub should_aggregate: bool,
+}
+
+impl UpstreamDifficultyConfig {
+    pub fn new(
+        channel_diff_update_interval: u32,
+        channel_nominal_hashrate: f32,
+        timestamp_of_last_update: u64,
+        should_aggregate: bool,
+    ) -> Self {
+        Self {
+            channel_diff_update_interval,
+            channel_nominal_hashrate,
+            timestamp_of_last_update,
+            should_aggregate,
+        }
+    }
 }
